backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pdfminer.six
openai
